feat(browser): rehydrate terminal panels on reload + extract registry driver
Spawn-time terminal-ui-launch events are one-shot and never replayed, so a Cmd+R
reload (which the browser user does often) or a fresh tab opening a project with
running agents showed no terminal panels even though the agents were alive in
tmux. browser terminal.rehydrate was a no-op.

- Implement browser terminal.rehydrate: pull authoritative records via the
  getTerminalRecords RPC and re-launch a panel for each live terminal through the
  ui:call seam — the browser parity of Electron's rehydrateTerminalPanels.
- Hoist the pure selector selectTerminalsToRehydrate (+ its black-box test) out of
  the Electron-only module (which also imports uiAPI) into shared shell/agent/terminals
  so both edges use one copy without dragging Electron deps into the browser bundle.
- Extract the browser terminal-registry UI driver (live subscription + rehydrate)
  into browserTerminalRegistry.ts, keeping browserRuntime.ts under the size limit.

Verified live: spawn renders the panel; reloading the tab restores every live
terminal panel (28 agents re-rendered post-reload in testing).
